+++ package.json
@@ -19,7 +19,8 @@
     },
     "homepage": "https://github.com/openaddresses/batch-alarms#readme",
     "dependencies": {
-        "@mapbox/cloudfriend": "^6.0.0"
+        "@mapbox/cloudfriend": "^6.0.0",
+        "documentation": "^14.0.1"
     },
     "devDependencies": {
         "eslint": "^8.0.1",
